Update README.md
Jaws is an American suspense and horror film directed by Steven Spielberg and released in 197512. It is considered the first summer blockbuster1. The film is based on the 1974 novel of the same name by author Peter Benchley1. It tells the story of a great white shark that terrorizes a beach resort town1. The protagonist, Police Chief Martin Brody, tries to protect beachgoers from the shark by closing the beach2.